+++ plugins/bizupkeep-workflow/includes/Workflows/CompanyAmendment/CompanyAmendmentService.php
@@ -6,6 +6,7 @@ namespace BizHub\Workflow\Workflows\CompanyAmendment;
 
 use BizHub\Companies\Contracts\CompanyServiceInterface;
 use BizHub\Workflow\Contracts\WorkflowEngineInterface;
+use BizHub\Workflow\Contracts\WorkflowTypeServiceInterface;
 use BizHub\Workflow\DTO\CreateWorkflowCommand;
 use BizHub\Workflow\DTO\RollbackWorkflowCommand;
 use BizHub\Workflow\DTO\TransitionWorkflowCommand;
@@ -20,7 +21,7 @@ use BizHub\Workflow\Exceptions\WorkflowNotFoundException;
  *
  * @package BizHub\Workflow\Workflows\CompanyAmendment
  */
-final class CompanyAmendmentService
+final class CompanyAmendmentService implements WorkflowTypeServiceInterface
 {
     /**
      * @var array<int,string>