+++ plugins/bizupkeep-workflow/includes/Workflows/CompanyRegistration/CompanyRegistrationService.php
@@ -6,6 +6,7 @@ namespace BizHub\Workflow\Workflows\CompanyRegistration;
 
 use BizHub\Companies\Contracts\CompanyServiceInterface;
 use BizHub\Workflow\Contracts\WorkflowEngineInterface;
+use BizHub\Workflow\Contracts\WorkflowTypeServiceInterface;
 use BizHub\Workflow\DTO\CreateWorkflowCommand;
 use BizHub\Workflow\DTO\RollbackWorkflowCommand;
 use BizHub\Workflow\DTO\TransitionWorkflowCommand;
@@ -26,7 +27,7 @@ use BizHub\Workflow\Exceptions\WorkflowNotFoundException;
  *
  * @package BizHub\Workflow\Workflows\CompanyRegistration
  */
-final class CompanyRegistrationService
+final class CompanyRegistrationService implements WorkflowTypeServiceInterface
 {
     /**
      * @var array<int,string>